src/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg[binary]==3.2.6
redis==5.2.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pydantic==2.11.1
pydantic-settings==2.8.1
email-validator==2.2.0
httpx==0.28.1
pytest==8.3.5
pytest-asyncio==0.26.0
aiosqlite==0.21.0
structlog==25.2.0
